Added make to list of packages installed
I use this project to mount my $GOPATH, so that I may test projects
which only work under Linux.  The ability to run `make` to test these
projects is necessary.  Also, seems to fall in line with the goal of
this project.  One would create a new go project, and should provide
a Makefile to execute tests.